PDF: collect atoms with a collector of their own
